moved schemas to its own folder
